PR 12 Fix compilation error after merge from base branch.

+++ flink/src/test/java/io/delta/flink/source/DeltaSourceContinuousExecutionITCaseTest.java
@@ -270,7 +270,7 @@ public class DeltaSourceContinuousExecutionITCaseTest extends DeltaSourceITBase 
         DeltaSourceConfiguration sourceConfiguration = source.getSourceConfiguration();
         // Main thread waits some time. To stay on a safe side, we wait doubled time of update
         // check interval. So source will have time to check table twice for updates.
-        int testTimeout =
+        long testTimeout =
             sourceConfiguration.getValue(DeltaSourceOptions.UPDATE_CHECK_INTERVAL) * 2;
         List<RowData> results = dataFuture.get(testTimeout, TimeUnit.MILLISECONDS);
